requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
Deprecated==1.2.14
Flask==3.0.1
Flask-Cors==4.0.0
fonttools==4.47.2
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mypy-extensions==1.0.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
Pint==0.23
platformdirs==4.2.0
pooch==1.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyvista==0.43.3
requests==2.31.0
scipy==1.12.0
scooby==0.9.2
six==1.16.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
vtk==9.3.0
Werkzeug==3.0.1
wrapt==1.16.0
